backend/requirements.txt
fastapi[standard]==0.115.12
pydantic==2.7.0
pydantic-settings==2.6.0
pydantic[email]==2.7.0
uvicorn[standard]==0.32.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.3.4
httpx==0.28.1
pytest-anyio==0.7.0
